+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+- Refactor PPT handling more:
+  1. Per profile, per-ac/dc
+  2. Do not apply unless group is enabled
+  3. Better reset/disable handling
+  4. Selecting a profile defaults PPT to off/disabled
+
 ## [v6.1.0-rc6]
 
 ### Changed